requirements.txt
openai==0.27.0
fastapi
black
isort
websockets
pydantic
langchain==0.0.126
uvicorn
jinja2
faiss-cpu
bs4
unstructured
libmagic
tiktoken
python-dotenv==1.0.0

